+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All the log of changes on the project/repository
 - Create a `serviceWorker` action and add it to the `cli build`
 - Styles and scripts dir constants
 - Refactor title to a function global `get_page_title`
+- Create a `create` helper function
+  - Implement `cli component` action
+  - Implement `cli action` action
 
 ### Modified
 - Contact `<img>` loaded directly